Adds multiple lock mode support  to the B+Tree.
