+++ src/Playground/Benchmark/BenchmarkGroups.cs
@@ -8,6 +8,12 @@ public static class BenchmarkGroups
         1_000_000,
         2_000_000,
         3_000_000,
+        4_000_000,
+        5_000_000,
+        6_000_000,
+        7_000_000,
+        8_000_000,
+        9_000_000,
         10_000_000
     };
