+++ src/Playground/InMemoryTreeBenchmark/RandomMidSizeInserts.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Tenray.ZoneTree.Collections;
-using Tenray.ZoneTree.Comparers;
+﻿using Tenray.ZoneTree.Collections;
+using Tenray.ZoneTree.Collections.BplusTree;
 
 namespace Playground.InMemoryTreeBenchmark;
