+++ src/ZoneTree.UnitTests/BplusTreeThreadSafetyTests.cs
@@ -1,4 +1,5 @@
 ﻿using Tenray.ZoneTree.Collections;
+using Tenray.ZoneTree.Collections.BplusTree;
 using Tenray.ZoneTree.Comparers;
 
 namespace Tenray.ZoneTree.UnitTests;
